third_party/blink/web_tests/virtual/web-components-v0-disabled/shadow-dom/css-focus-pseudo-match-shadow-host1-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
:focus and shadow host without delegatesFocus for crbug/479050
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
